Fixed the elif that wasn't working
I fixed the elif that wasn't working in the Guess_Number Game. So I had to separate the elif statement to make them 2, got the idea from this statement, "Python is much like natural language but the interpreter cannot fill in what human listeners can. 'a and b in c' must be written out as 'a in c and b in c'." that I met on stack overflow (https://stackoverflow.com/questions/36070252/elif-statement-combined-with-and-or-or-not-working)